Add files via upload
This is a weatherapp  application that uses an API to receive the current weather of any city in the world.
You will need an API key in other to run this application. 
You can sign up for one at www.openweathermap.org.

Please installl npm yargs into a folder. This is an Asynchronous Programming.
Uses can run with command node app.js -l {city goes here}